Check SG terms endpoint first on search before doing keyword
This looks to see if we can get a good match on the Scigraph Terms
endpoint before just launching into a keyword search. If we get a single
match on terms, we redirect to that wiki page. Otherwise merge in the
result set.

Also some better 404 rendering. If a wiki page isn't found, we redirect
to the 404 page.