Change redirection behavior on Search results page
This improves the behavior on the search results page, primarily
focusing on redirection. If a search query returns a single result from
the Terms scigraph endpoint, we redirect to that wiki entry on default.

However, this feature can be turned off with a "redirect=false" param
passed into the query string. This is used in the "Search This Term"
link, which is included in the Wiki title. This allows for users to see
related results w/out being automatically redirected.

https://pm.incf.org/browse/KNOW-24
https://pm.incf.org/browse/KNOW-26